Add material implementations for Lambertian, Metal, Blinn-Phong, and Dielectric; update material module structure
