+++ src/lib.rs
@@ -9,16 +9,16 @@ pub mod light;
 pub mod material;
 pub mod ray;
 pub mod sphere;
-pub mod vec3;
 pub mod tracer;
+pub mod vec3;
 pub mod world;
 
 pub use buffer::Buffer;
 pub use camera::Camera;
 pub use color::Color;
 pub use hittable_list::HittableList;
-pub use ray::Ray;
-pub use vec3::{Vec3, Point3};
-pub use material::*;
 pub use light::{Light, LightList};
+pub use material::*;
+pub use ray::Ray;
+pub use vec3::{Point3, Vec3};
 pub use world::random_scene;
