+++ src/material/dielectric.rs
@@ -1,0 +1,58 @@
+use crate::color::Color;
+use crate::common;
+use crate::hittable::HitRecord;
+use crate::material::Material;
+use crate::ray::Ray;
+use crate::vec3;
+
+pub struct Dielectric {
+    ir: f32, // Index of refraction
+}
+
+impl Dielectric {
+    pub fn new(index_of_refraction: f32) -> Dielectric {
+        Dielectric {
+            ir: index_of_refraction,
+        }
+    }
+
+    fn reflectance(cosine: f32, ref_idx: f32) -> f32 {
+        // Use Schlick's approximation for reflectance
+        let mut r0 = (1.0 - ref_idx) / (1.0 + ref_idx);
+        r0 = r0 * r0;
+        r0 + (1.0 - r0) * f32::powf(1.0 - cosine, 5.0)
+    }
+}
+
+impl Material for Dielectric {
+    fn scatter(
+        &self,
+        r_in: &Ray,
+        rec: &HitRecord,
+        attenuation: &mut Color,
+        scattered: &mut Ray,
+    ) -> bool {
+        let refraction_ratio = if rec.front_face {
+            1.0 / self.ir
+        } else {
+            self.ir
+        };
+
+        let unit_direction = vec3::unit_vector(r_in.direction());
+        let cos_theta = f32::min(vec3::dot(-unit_direction, rec.normal), 1.0);
+        let sin_theta = f32::sqrt(1.0 - cos_theta * cos_theta);
+
+        let cannot_refract = refraction_ratio * sin_theta > 1.0;
+        let direction = if cannot_refract
+            || Self::reflectance(cos_theta, refraction_ratio) > common::random()
+        {
+            vec3::reflect(unit_direction, rec.normal)
+        } else {
+            vec3::refract(unit_direction, rec.normal, refraction_ratio)
+        };
+
+        *attenuation = Color::new(1.0, 1.0, 1.0);
+        *scattered = Ray::new(rec.p, direction);
+        true
+    }
+}
