+++ src/material/lambert.rs
@@ -1,0 +1,36 @@
+use crate::color::Color;
+use crate::hittable::HitRecord;
+use crate::material::Material;
+use crate::ray::Ray;
+use crate::vec3;
+
+pub struct Lambertian {
+    albedo: Color,
+}
+
+impl Lambertian {
+    pub fn new(a: Color) -> Lambertian {
+        Lambertian { albedo: a }
+    }
+}
+
+impl Material for Lambertian {
+    fn scatter(
+        &self,
+        _r_in: &Ray,
+        rec: &HitRecord,
+        attenuation: &mut Color,
+        scattered: &mut Ray,
+    ) -> bool {
+        let mut scatter_direction = rec.normal + vec3::random_unit_vector();
+
+        // Catch degenerate scatter direction
+        if scatter_direction.near_zero() {
+            scatter_direction = rec.normal;
+        }
+
+        *attenuation = self.albedo;
+        *scattered = Ray::new(rec.p, scatter_direction);
+        true
+    }
+}
