+++ src/material/metal.rs
@@ -1,0 +1,35 @@
+use crate::color::Color;
+use crate::hittable::HitRecord;
+use crate::material::Material;
+use crate::ray::Ray;
+use crate::vec3;
+
+pub struct Metal {
+    albedo: Color,
+    fuzz: f32,
+}
+
+impl Metal {
+    pub fn new(a: Color, f: f32) -> Metal {
+        Metal {
+            albedo: a,
+            fuzz: if f < 1.0 { f } else { 1.0 },
+        }
+    }
+}
+
+impl Material for Metal {
+    fn scatter(
+        &self,
+        r_in: &Ray,
+        rec: &HitRecord,
+        attenuation: &mut Color,
+        scattered: &mut Ray,
+    ) -> bool {
+        let reflected = vec3::reflect(vec3::unit_vector(r_in.direction()), rec.normal);
+
+        *attenuation = self.albedo;
+        *scattered = Ray::new(rec.p, reflected + self.fuzz * vec3::random_in_unit_sphere());
+        vec3::dot(scattered.direction(), rec.normal) > 0.0
+    }
+}
